+++ click_extra/__init__.py
@@ -242,7 +242,7 @@ __all__ = [  # noqa: F405
 """
 
 
-__version__ = "7.6.3"
+__version__ = "7.6.4.dev0"
 
 
 def __getattr__(name: str) -> object: